FEAT: repo domain db 저장.. 추가 기능 구현 등..
