+++ src/main/java/com/Teletubbies/Apollo/auth/domain/Repo.java
@@ -1,13 +1,12 @@
 package com.Teletubbies.Apollo.auth.domain;
 
 import com.Teletubbies.Apollo.auth.dto.RepoInfoResponse;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
+import jakarta.persistence.*;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
+@Getter
 @NoArgsConstructor
 public class Repo {
     @Id
@@ -15,7 +14,9 @@ public class Repo {
     private Long id;
     private String repoName;
     private String repoUrl;
+    private String ownerLogin;
     public Repo(RepoInfoResponse response){
+        this.ownerLogin = response.getUserLogin();
         this.repoName = response.getRepoName();
         this.repoUrl = response.getRepoURL();
     }
